+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 * 25/01/12
+    - 0.1.0 (utils/email.py) <br>
+        css embedding style to inline style <br>
+
+* 25/01/12
     - 0.0.9 (utils/db.py) <br>
         apply save_news_to_db <br>